pyproject.toml
[project]
name = "polis-calc"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiocache[redis]>=0.12.3",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi[all]>=0.116.1",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.43",
]

[dependency-groups]
dev = [
    "ruff>=0.12.10",
]

[tool.ruff]
line-length = 100
exclude = [".venv", "venv", "migrations", "alembic"]
